// src/utils/enums.ts
export enum EmployeeDayStatus {
  NOT_STARTED = "NOT_STARTED",
  STARTED = "STARTED",
  CHECKED_IN = "CHECKED_IN",
  CHECKED_OUT = "CHECKED_OUT",
  DAY_COMPLETED = "DAY_COMPLETED",
  CHECKED_IN_NO_CHECKOUT = "CHECKED_IN_NO_CHECKOUT",
  OUT_SICK = "OUT_SICK",
  EARLY_OUT = "EARLY_OUT",
  ABSENT = "ABSENT",
  TEMPORARILY_UNAVAILABLE = "TEMPORARILY_UNAVAILABLE",
  PERMISSION_GRANTED = "PERMISSION_GRANTED",
  ON_BREAK = "ON_BREAK",
}
export enum Status {
  IDLE = "idle",
  LOADING = "loading",
  SUCCEEDED = "succeeded",
  FAILED = "failed",
}
export enum Roles {
  EMPLOYEE = "EMPLOYEE",
  ADMIN = "ADMIN",
  SUPER_ADMIN = "SUPER_ADMIN",
}